dougslldataplatform/pyproject.toml
[tool.poetry]
name = "dougslldataplatform"
version = "0.1.0"
description = ""
authors = ["Leal <douglas.sleal@outlook.com>"]
readme = "README.md"

[tool.poetry.scripts]
dougslldataplatform = "dougslldataplatform.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.7.1"
typer = "^0.12.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.12.2"
databricks-sdk = "^0.27.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.21"
mkdocstrings = "^0.25.0"
mkdocstrings-python = "^1.10.0"
mkdocs-macros-plugin = "^1.0.5"
jinja2 = "^3.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=dougslldataplatform -vv"
post_test = "coverage html"
